Rename tool to vs, renamed list to where
And simplify it entirely so `vs where` is a straight passthrough to `vswhere.exe`, including argument formatting and all. This makes it far simpler to document and keep in sync (nothing to sync!), even if it comes at the slight expense of missing consistency with the way Mono.Options exposes arguments.