Add Run command
The new `run` command is also the default command unless another
one or `-h` is provided. This makes it trivial to use it as a
shorthand for running VS from anywhere simply with `vs`.

Additional options to specify which instance to run are provided,
which also are unified with the `install` and `where` commands,
so you can, for example, run a VS with the .NET Core workload
installed by using `vs +core`, for example.

The parsing of the `--sku=[edition]` is also shared across `run`,
`install` and `where` so it can be used uniformly with all commands,
such as `vs where --sku=ent`.