Improve default install path probing
By default, the installer will choose a target folder named after the SKU.
If we find an existing SKU, we try `Preview` or `IntPreview` depending
on `--preview` or `--dogfood` switches. If we still find an existing installation,
we try `Pre` or `Int` + SKU, which makes it almost impossible to have dupes.