Add support for `vs config`
This opens the VS configuration directory under `%LocalAppData%\Microsoft\VisualStudio\[MAJOR.0]_[ID][Exp]`.

Allows passing `-exp` to open the experimental folder instead of the regular instance.

Added option to pass in a verb to use for the channel switch, so that it doesn't read "Install ..." on all
commands using the `VisualStudioOptions`.