Add back run as the default command, improve its selector
Add channel and experimental instance selection as part of the reusable
VS options.

Make run the default command, as it was before. We'll still render
usage if `vs -?` is used or if `vs` with no arguments is run too.

Added return value to main function just in case consumers want to determine
if the command run successfully or not.