Consider Run default command for saved aliases too
Since `run` is the default command, the alias could be saved without
an explicit command. When reading the alias in that case, we couldn't
find a matching factory and ignored the entire alias arguments.

This commit adds a fallback so that if the first alias argument cannot
be interpreted as a command, we also apply the default behavior for the
non-save scenario and try it as a `Run` command instead.