Reintroduce --version option in a vswhere-compatible way
We had disabled this in  9093d5a6d2d5e80d70135e61c8b98e898614fd9a because it previously interfered with the ability to run vs where with a version specifier.

We now instead check that the version is defined as the *first* argument specified, which covers that and makes it so that only `vs --version` will show it.

Fixes #72
Fixes #73