Improve handling of update/modify scenarios
We previously hardcoded the major VS version used to build the channel bootstrapper, but this was only necessary for Install scenarios, where we don't know already the channel/installation we're dealing with.

So for update/modify, we pass down the selected instance (or default selected) major version, so we don't select the wrong major version inadvertently.

In addition, for Install, we have a smarter heuristics now that will attempt to load the latest release bootstrapper major version from the official docs on CLI installs for VS, and defaulting to hardcoded "17" for now in the unlikely scenario where we can't reach the page for some reason.

Fixes #123